QueryService split data grpc methods and public interface
