+++ src/Ydb.Sdk/src/Services/Query/QueryClient.cs
@@ -1,297 +1,9 @@
 using Microsoft.Extensions.Logging;
-using Ydb.Query;
-using Ydb.Query.V1;
 using Ydb.Sdk.Client;
 using Ydb.Sdk.Services.Sessions;
 using Ydb.Sdk.Value;
 
 namespace Ydb.Sdk.Services.Query;
-
-public enum ExecMode
-{
-    Unspecified = 0,
-    Parse = 10,
-    Validate = 20,
-    Explain = 30,
-
-    // reserved 40; // EXEC_MODE_PREPARE
-    Execute = 50
-}
-
-public enum Syntax
-{
-    Unspecified = 0,
-
-    /// <summary>
-    /// YQL
-    /// </summary>
-    YqlV1 = 1,
-
-    /// <summary>
-    /// PostgresQL
-    /// </summary>
-    Pg = 2
-}
-
-public enum StatsMode
-{
-    Unspecified = 0,
-
-    /// <summary>
-    /// Stats collection is disabled
-    /// </summary>
-    None = 10,
-
-    /// <summary>
-    /// Aggregated stats of reads, updates and deletes per table
-    /// </summary>
-    Basic = 20,
-
-    /// <summary>
-    /// Add execution stats and plan on top of STATS_MODE_BASIC
-    /// </summary>
-    Full = 30,
-
-    /// <summary>
-    /// Detailed execution stats including stats for individual tasks and channels
-    /// </summary>
-    Profile = 40
-}
-
-public class ExecuteQuerySettings : RequestSettings
-{
-    public ExecMode ExecMode { get; set; } = ExecMode.Execute;
-    public Syntax Syntax { get; set; }
-
-    public StatsMode StatsMode { get; set; }
-}
-
-public class CreateSessionSettings : RequestSettings
-{
-}
-
-public class DeleteSessionSettings : RequestSettings
-{
-}
-
-public class AttachSessionSettings : RequestSettings
-{
-}
-
-public class BeginTransactionSettings : RequestSettings
-{
-}
-
-public class CommitTransactionSettings : RequestSettings
-{
-}
-
-public class RollbackTransactionSettings : RequestSettings
-{
-}
-
-public class CreateSessionResponse : ResponseWithResultBase<CreateSessionResponse.ResultData>
-{
-    internal CreateSessionResponse(Status status, ResultData? result = null)
-        : base(status, result)
-    {
-    }
-
-    public class ResultData
-    {
-        private ResultData(Session session)
-        {
-            Session = session;
-        }
-
-        public Session Session { get; }
-
-        internal static ResultData FromProto(Ydb.Query.CreateSessionResponse resultProto, Driver driver,
-            string endpoint)
-        {
-            var session = new Session(
-                driver: driver,
-                sessionPool: null,
-                id: resultProto.SessionId,
-                nodeId: resultProto.NodeId,
-                endpoint: endpoint);
-
-            return new ResultData(
-                session: session
-            );
-        }
-    }
-}
-
-public class DeleteSessionResponse : ResponseBase
-{
-    internal DeleteSessionResponse(Status status) : base(status)
-    {
-    }
-
-    private DeleteSessionResponse(Ydb.Query.DeleteSessionResponse proto)
-        : base(Status.FromProto(proto.Status, proto.Issues))
-    {
-    }
-
-    internal static DeleteSessionResponse FromProto(Ydb.Query.DeleteSessionResponse proto)
-    {
-        return new DeleteSessionResponse(proto);
-    }
-}
-
-public class SessionState : ResponseBase
-{
-    internal SessionState(Status status) : base(status)
-    {
-    }
-
-    private SessionState(Ydb.Query.SessionState proto)
-        : base(Status.FromProto(proto.Status, proto.Issues))
-    {
-    }
-
-    internal static SessionState FromProto(Ydb.Query.SessionState proto)
-    {
-        return new SessionState(proto);
-    }
-}
-
-public class SessionStateStream : StreamResponse<Ydb.Query.SessionState, SessionState>
-{
-    internal SessionStateStream(Driver.StreamIterator<Ydb.Query.SessionState> iterator) : base(iterator)
-    {
-    }
-
-    protected override SessionState MakeResponse(Ydb.Query.SessionState protoResponse)
-    {
-        return SessionState.FromProto(protoResponse);
-    }
-
-    protected override SessionState MakeResponse(Status status)
-    {
-        return new SessionState(status);
-    }
-}
-
-public class ExecuteQueryResponsePart : ResponseBase
-{
-    public long? ResultSetIndex;
-    public Value.ResultSet? ResultSet;
-
-    internal ExecuteQueryResponsePart(Status status) : base(status)
-    {
-    }
-
-    private ExecuteQueryResponsePart(Ydb.Query.ExecuteQueryResponsePart proto)
-        : base(Status.FromProto(proto.Status, proto.Issues))
-    {
-        ResultSetIndex = proto.ResultSetIndex;
-        if (proto.ResultSet is not null)
-        {
-            ResultSet = Value.ResultSet.FromProto(proto.ResultSet);
-        }
-    }
-
-    internal static ExecuteQueryResponsePart FromProto(Ydb.Query.ExecuteQueryResponsePart proto)
-    {
-        return new ExecuteQueryResponsePart(proto);
-    }
-}
-
-public class ExecuteQueryStream : StreamResponse<Ydb.Query.ExecuteQueryResponsePart, ExecuteQueryResponsePart>
-    , IAsyncEnumerable<ExecuteQueryResponsePart>
-{
-    internal ExecuteQueryStream(Driver.StreamIterator<Ydb.Query.ExecuteQueryResponsePart> iterator) : base(iterator)
-    {
-    }
-
-    public new async Task<bool> Next()
-    {
-        var isNext = await base.Next();
-        if (isNext)
-        {
-            Response.EnsureSuccess();
-        }
-
-        return isNext;
-    }
-
-    protected override ExecuteQueryResponsePart MakeResponse(Ydb.Query.ExecuteQueryResponsePart protoResponse)
-    {
-        return ExecuteQueryResponsePart.FromProto(protoResponse);
-    }
-
-    protected override ExecuteQueryResponsePart MakeResponse(Status status)
-    {
-        return new ExecuteQueryResponsePart(status);
-    }
-
-    public async IAsyncEnumerator<ExecuteQueryResponsePart> GetAsyncEnumerator(
-        CancellationToken cancellationToken = new CancellationToken())
-    {
-        while (await Next())
-        {
-            yield return Response;
-        }
-    }
-}
-
-public class BeginTransactionResponse : ResponseBase
-{
-    internal BeginTransactionResponse(Status status) : base(status)
-    {
-    }
-
-    public Tx Tx { get; } = new();
-
-    private BeginTransactionResponse(Ydb.Query.BeginTransactionResponse proto) : base(
-        Status.FromProto(proto.Status, proto.Issues))
-    {
-        var txId = proto.TxMeta.Id;
-        var tx = new Tx(new TransactionControl { TxId = txId });
-        Tx = tx;
-    }
-
-    internal static BeginTransactionResponse FromProto(Ydb.Query.BeginTransactionResponse proto)
-    {
-        return new BeginTransactionResponse(proto);
-    }
-}
-
-public class CommitTransactionResponse : ResponseBase
-{
-    internal CommitTransactionResponse(Status status) : base(status)
-    {
-    }
-
-    private CommitTransactionResponse(Ydb.Query.CommitTransactionResponse proto) : base(
-        Status.FromProto(proto.Status, proto.Issues))
-    {
-    }
-
-    internal static CommitTransactionResponse FromProto(Ydb.Query.CommitTransactionResponse proto)
-    {
-        return new CommitTransactionResponse(proto);
-    }
-}
-
-public class RollbackTransactionResponse : ResponseBase
-{
-    internal RollbackTransactionResponse(Status status) : base(status)
-    {
-    }
-
-    private RollbackTransactionResponse(Ydb.Query.RollbackTransactionResponse proto) : base(
-        Status.FromProto(proto.Status, proto.Issues))
-    {
-    }
-
-    internal static RollbackTransactionResponse FromProto(Ydb.Query.RollbackTransactionResponse proto)
-    {
-        return new RollbackTransactionResponse(proto);
-    }
-}
 
 public class QueryClientConfig
 {
@@ -304,9 +16,7 @@ public class QueryClientConfig
     }
 }
 
-public class QueryClient :
-    ClientBase,
-    IDisposable
+public class QueryClient : QueryClientGrpc, IDisposable
 {
     private readonly ISessionPool<Session> _sessionPool;
     private readonly ILogger _logger;
@@ -328,173 +38,6 @@ public class QueryClient :
         _sessionPool = sessionPool;
     }
 
-    internal async Task<CreateSessionResponse> CreateSession(CreateSessionSettings? settings = null)
-    {
-        settings ??= new CreateSessionSettings();
-        var request = new CreateSessionRequest();
-
-        try
-        {
-            var response = await Driver.UnaryCall(
-                method: QueryService.CreateSessionMethod,
-                request: request,
-                settings: settings);
-
-            var status = Status.FromProto(response.Data.Status, response.Data.Issues);
-
-            CreateSessionResponse.ResultData? result = null;
-
-            if (status.IsSuccess)
-            {
-                result = CreateSessionResponse.ResultData.FromProto(response.Data, Driver, response.UsedEndpoint);
-            }
-
-            return new CreateSessionResponse(status, result);
-        }
-        catch (Driver.TransportException e)
-        {
-            return new CreateSessionResponse(e.Status);
-        }
-    }
-
-    internal async Task<DeleteSessionResponse> DeleteSession(string sessionId, DeleteSessionSettings? settings = null)
-    {
-        settings ??= new DeleteSessionSettings();
-        var request = new DeleteSessionRequest
-        {
-            SessionId = sessionId
-        };
-
-        try
-        {
-            var response = await Driver.UnaryCall(
-                method: QueryService.DeleteSessionMethod,
-                request: request,
-                settings: settings);
-
-
-            return DeleteSessionResponse.FromProto(response.Data);
-        }
-        catch (Driver.TransportException e)
-        {
-            return new DeleteSessionResponse(e.Status);
-        }
-    }
-
-    internal SessionStateStream AttachSession(string sessionId, AttachSessionSettings? settings = null)
-    {
-        settings ??= new AttachSessionSettings { TransportTimeout = TimeSpan.FromDays(1) };
-
-        var request = new AttachSessionRequest { SessionId = sessionId };
-
-        var streamIterator = Driver.StreamCall(
-            method: QueryService.AttachSessionMethod,
-            request: request,
-            settings: settings
-        );
-        return new SessionStateStream(streamIterator);
-    }
-
-    private async Task<BeginTransactionResponse> BeginTransaction(
-        string sessionId,
-        Tx tx,
-        BeginTransactionSettings? settings = null)
-    {
-        settings ??= new BeginTransactionSettings();
-
-        var request = new BeginTransactionRequest { SessionId = sessionId, TxSettings = tx.ToProto().BeginTx };
-        try
-        {
-            var response = await Driver.UnaryCall(
-                QueryService.BeginTransactionMethod,
-                request: request,
-                settings: settings
-            );
-            return BeginTransactionResponse.FromProto(response.Data);
-        }
-        catch (Driver.TransportException e)
-        {
-            return new BeginTransactionResponse(e.Status);
-        }
-    }
-
-    private async Task<CommitTransactionResponse> CommitTransaction(
-        string sessionId,
-        Tx tx,
-        CommitTransactionSettings? settings = null)
-    {
-        settings ??= new CommitTransactionSettings();
-
-        var request = new CommitTransactionRequest { SessionId = sessionId, TxId = tx.TxId };
-
-        try
-        {
-            var response = await Driver.UnaryCall(
-                QueryService.CommitTransactionMethod,
-                request: request,
-                settings: settings
-            );
-            return CommitTransactionResponse.FromProto(response.Data);
-        }
-        catch (Driver.TransportException e)
-        {
-            return new CommitTransactionResponse(e.Status);
-        }
-    }
-
-    private async Task<RollbackTransactionResponse> RollbackTransaction(
-        string sessionId,
-        Tx tx,
-        RollbackTransactionSettings? settings = null)
-    {
-        settings ??= new RollbackTransactionSettings();
-
-        var request = new RollbackTransactionRequest { SessionId = sessionId, TxId = tx.TxId };
-        try
-        {
-            var response = await Driver.UnaryCall(
-                QueryService.RollbackTransactionMethod,
-                request: request,
-                settings: settings
-            );
-            return RollbackTransactionResponse.FromProto(response.Data);
-        }
-        catch (Driver.TransportException e)
-        {
-            return new RollbackTransactionResponse(e.Status);
-        }
-    }
-
-
-    internal ExecuteQueryStream ExecuteQuery(
-        string sessionId,
-        string queryString,
-        Tx tx,
-        IReadOnlyDictionary<string, YdbValue>? parameters,
-        ExecuteQuerySettings? settings = null)
-    {
-        settings ??= new ExecuteQuerySettings();
-        parameters ??= new Dictionary<string, YdbValue>();
-
-        var request = new ExecuteQueryRequest
-        {
-            SessionId = sessionId,
-            ExecMode = (Ydb.Query.ExecMode)settings.ExecMode,
-            TxControl = tx.ToProto(),
-            QueryContent = new QueryContent { Syntax = (Ydb.Query.Syntax)settings.Syntax, Text = queryString },
-            StatsMode = (Ydb.Query.StatsMode)settings.StatsMode
-        };
-
-        request.Parameters.Add(parameters.ToDictionary(p => p.Key, p => p.Value.GetProto()));
-
-        var streamIterator = Driver.StreamCall(
-            method: QueryService.ExecuteQueryMethod,
-            request: request,
-            settings: settings);
-
-        return new ExecuteQueryStream(streamIterator);
-    }
-
     private async Task<IResponse> ExecOnSession(
         Func<Session, Task<IResponse>> func,
         RetrySettings? retrySettings = null
@@ -509,12 +52,14 @@ public class QueryClient :
         return await sessionPool.ExecOnSession(func, retrySettings);
     }
 
-    internal static async Task EmptyStreamReadFunc(ExecuteQueryStream stream)
+    internal static async Task<None> EmptyStreamReadFunc(ExecuteQueryStream stream)
     {
         while (await stream.Next())
         {
             stream.Response.EnsureSuccess();
         }
+
+        return None.Instance;
     }
 
     public async Task<QueryResponseWithResult<T>> Query<T>(
@@ -567,11 +112,7 @@ public class QueryClient :
         var response = await Query<None>(
             queryString,
             parameters,
-            async session =>
-            {
-                await EmptyStreamReadFunc(session);
-                return None.Instance;
-            },
+            async session => await EmptyStreamReadFunc(session),
             txModeSettings,
             executeQuerySettings,
             retrySettings);
