+++ src/Ydb.Sdk/src/Services/Query/Session.cs
