+++ src/Ydb.Sdk/src/Services/Sessions/SessionBase.cs
